Warnings: Fix floating point conversion warnings

+++ tests/gui/realtime/rt_tests.cpp
@@ -17,14 +17,14 @@ TEST(random_amount_y_data, real_time) {
 
   GET_TIMER_CB = [=](const int dt_ms) {
     static std::vector<std::vector<float>> y_test_data;
-    size_t points_count = 2 + (size_t)(((float)rand() / RAND_MAX) * 1000.0);
-    size_t curve_count = 1 + (size_t)(((float)rand() / RAND_MAX) * 10.0);
+    size_t points_count = 2 + (size_t)((float(rand()) / float(RAND_MAX)) * 1000.0);
+    size_t curve_count = 1 + (size_t)((float(rand()) / float(RAND_MAX)) * 10.0);
 
     y_test_data.resize(curve_count);
     for (size_t i = 0; i < curve_count; ++i) {
       y_test_data[i].resize(points_count);
       for (size_t j = 0; j < points_count; ++j) {
-        y_test_data[i][j] = (float)rand() / RAND_MAX;
+        y_test_data[i][j] = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
       }
     }
     PLOT_Y({y_test_data});
@@ -39,7 +39,7 @@ TEST(random_y_values, real_time) {
     static std::vector<float> y_test_data(10);
 
     for (auto& y : y_test_data) {
-      y = (float)rand() / RAND_MAX;
+      y = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
     }
 
     PLOT_Y({y_test_data});
@@ -58,7 +58,7 @@ TEST(real_time_plot_function, real_time) {
     static std::vector<float> y_test_data(10);
 
     for (auto& y : y_test_data) {
-      y = (float)rand() / RAND_MAX;
+      y = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
     }
 
     GET_PLOT->plotUpdateYOnly({y_test_data});
@@ -103,8 +103,8 @@ TEST(random_yx_values, real_time) {
 
   GET_TIMER_CB = [=](const int dt_ms) {
     static std::vector<std::vector<float>> x_test_data, y_test_data;
-    size_t points_count = 2 + (size_t)(((float)rand() / RAND_MAX) * 1000.0);
-    size_t curve_count = 1 + (size_t)(((float)rand() / RAND_MAX) * 10.0);
+    size_t points_count = 2 + (size_t)((float(rand()) / float(RAND_MAX)) * 1000.0);
+    size_t curve_count = 1 + (size_t)((float(rand()) / float(RAND_MAX)) * 10.0);
 
     x_test_data.resize(curve_count);
     y_test_data.resize(curve_count);
@@ -113,7 +113,7 @@ TEST(random_yx_values, real_time) {
       y_test_data[i].resize(points_count);
       for (size_t j = 0; j < points_count; ++j) {
         x_test_data[i][j] = (float)j;  // Linear curve
-        y_test_data[i][j] = (float)rand() / RAND_MAX;
+        y_test_data[i][j] = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
       }
     }
     PLOT_XY({y_test_data}, {x_test_data});
